Revert "A"
This reverts commit 4bbbbf65d5dc743e618995a3a01250e13a5224b0.

+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/movement/speeds/ncp/OldNCPYPort.kt
@@ -31,7 +31,7 @@ class OldNCPBHop : SpeedMode("OldNCPBHop") {
                 mc.thePlayer.motionX -= sin(f) * 0.089;
                 mc.thePlayer.motionZ += cos(f) * 0.089;
             }else{
-                if (mc.thePlayer.fallDistance <= 1) mc.thePlayer.motionY = -1
+                if (mc.thePlayer.fallDistance <= (1).toDouble()) mc.thePlayer.motionY = (-1).toDouble()
             }
         } else {
             mc.thePlayer.motionX = 0.0
